requirements.txt
alembic==1.0.11
APScheduler==3.6.1
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
micro-service-authorization==0.0.2
micro-service-request-tool==0.0.8
pycodestyle==2.5.0
pycryptodomex==3.7.2
pyflakes==2.1.1
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
raven==6.10.0
redis==3.3.8
request-healthy==0.0.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.7
tzlocal==2.0.0
urllib3==1.25.3
Werkzeug==0.15.5
xmltodict==0.12.0
